chore: refresh docs and update tooling
