+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -7,15 +7,6 @@ assignees: swimmwatch
 
 ---
 
----
-name: Bug report
-about: Create a report to help us improve
-title: ''
-labels: bug
-assignees: swimmwatch
-
----
-
 <!--- Provide a general summary of the issue in the Title above -->
 
 ## Expected Behavior
@@ -38,11 +29,3 @@ assignees: swimmwatch
 ## Context (Environment)
 <!--- How has this issue affected you? What are you trying to accomplish? -->
 <!--- Providing context helps us come up with a solution that is most useful in the real world -->
-
-<!--- Provide a general summary of the issue in the Title above -->
-
-## Detailed Description
-<!--- Provide a detailed description of the change or addition you are proposing -->
-
-## Possible Implementation
-<!--- Not obligatory, but suggest an idea for implementing addition or change -->
