+++ CONTRIBUTING.md
@@ -1,17 +1,20 @@
 # Contributing
-We love your input! We want to make contributing to this project as easy and transparent as possible, whether it's:
+
+This project is in maintenance mode. We accept pull requests for bug fixes only.
+
+We want to make contributing to this project as easy and transparent as possible, whether it's:
 
 - Reporting a bug
 - Discussing the current state of the code
 - Submitting a fix
-- Proposing new features
-- Becoming a maintainer
 
-## We Develop with Github
-We use github to host code, to track issues and feature requests, as well as accept pull requests.
+## We Develop with GitHub
 
-## We Use [Github Flow](https://guides.github.com/introduction/flow/index.html), So All Code Changes Happen Through Pull Requests
-Pull requests are the best way to propose changes to the codebase (we use [Github Flow](https://guides.github.com/introduction/flow/index.html)). We actively welcome your pull requests:
+We use GitHub to host code, track issues, and accept pull requests.
+
+## We Use [GitHub Flow](https://guides.github.com/introduction/flow/index.html), So All Code Changes Happen Through Pull Requests
+
+Pull requests are the best way to propose bug fixes to the codebase. We actively welcome maintenance pull requests:
 
 1. Fork the repo and create your branch from `dev`.
 2. If you've added code that should be tested, add tests.
@@ -20,10 +23,12 @@ Pull requests are the best way to propose changes to the codebase (we use [Githu
 5. Make sure your code lints.
 6. Issue that pull request!
 
-## Report bugs using Github's [issues](https://github.com/swimmwatch/telegram-webapp-auth/issues)
-We use GitHub issues to track public bugs. Report a bug by [opening a new issue](); it's that easy!
+## Report Bugs Using GitHub's [Issues](https://github.com/swimmwatch/telegram-webapp-auth/issues)
 
-## Write bug reports with detail, background, and sample code
+We use GitHub issues to track public bugs. Report a bug by [opening a new issue](https://github.com/swimmwatch/telegram-webapp-auth/issues/new?template=bug_report.md).
+
+## Write Bug Reports with Detail, Background, and Sample Code
+
 [This is an example](http://stackoverflow.com/q/12488905/180626) of a bug report, and I think it's not a bad model.
 
 **Great Bug Reports** tend to have:
@@ -36,7 +41,6 @@ We use GitHub issues to track public bugs. Report a bug by [opening a new issue]
 - What actually happens
 - Notes (possibly including why you think this might be happening, or stuff you tried that didn't work)
 
-People *love* thorough bug reports. I'm not even kidding.
-
 ## References
+
 This document was adapted from the open-source contribution guidelines for [Facebook's Draft](https://github.com/facebook/draft-js/blob/a9316a723f9e918afde44dea68b5f9f39b7d9b00/CONTRIBUTING.md)
