+++ docs/guide/examples/django.md
@@ -1,66 +1,76 @@
-## Using with Django
-Let's create authorization middleware.
+# Django Integration
 
-Firstly, create variables in your `settings.py`:
+For Django applications, validate Telegram init data in middleware or in a dedicated authentication helper before resolving the current user.
+
+## Settings
+
+Generate the Telegram secret key from an environment-provided bot token:
+
 ```python
 from telegram_webapp_auth.auth import generate_secret_key
 
-# other settings
-
-TELEGRAM_BOT_TOKEN = env.str('TELEGRAM_BOT_TOKEN')
+TELEGRAM_BOT_TOKEN = env.str("TELEGRAM_BOT_TOKEN")
 TELEGRAM_SECRET_KEY = generate_secret_key(TELEGRAM_BOT_TOKEN)
 ```
 
-Then implement middleware:
+## Middleware
 
 ```python
+from datetime import timedelta
+
 from django.conf import settings
-from django.contrib.auth.models import User
+from django.contrib.auth import get_user_model
+from django.core.exceptions import PermissionDenied
 from django.http import HttpRequest
 from django.http import HttpResponse
+
 from telegram_webapp_auth.auth import TelegramAuthenticator
+from telegram_webapp_auth.errors import ExpiredInitDataError
 from telegram_webapp_auth.errors import InvalidInitDataError
 
 
-class TWAAuthorizationMiddleware:
+class TelegramWebAppAuthMiddleware:
     def __init__(self, get_response) -> None:
         self.get_response = get_response
-        self._telegram_authenticator = TelegramAuthenticator(settings.TELEGRAM_SECRET_KEY)
+        self.authenticator = TelegramAuthenticator(settings.TELEGRAM_SECRET_KEY)
 
     def __call__(self, request: HttpRequest) -> HttpResponse:
-        # Code to be executed for each request before
-        # the view (and later middleware) are called.
-        auth_cred = request.headers.get('Authorization')
+        auth_header = request.headers.get("Authorization")
+        if not auth_header:
+            raise PermissionDenied("Authorization header is required")
 
         try:
-            init_data = self._telegram_authenticator.validate(auth_cred)
-        except InvalidInitDataError:
-            # TODO: handle error
-            pass
-        
-        if not init_data.user:
-            # TODO: handle error
-            pass
+            init_data = self.authenticator.validate(
+                init_data=auth_header,
+                expr_in=timedelta(minutes=5),
+            )
+        except ExpiredInitDataError as exc:
+            raise PermissionDenied("Telegram init data has expired") from exc
+        except InvalidInitDataError as exc:
+            raise PermissionDenied("Telegram init data is invalid") from exc
 
-        current_user = User.objects.filter(tg_id=init_data.user.id).first()
-        if not current_user:
-            # TODO: handle error
-            pass
+        if init_data.user is None:
+            raise PermissionDenied("Telegram user data is required")
 
-        request.user = current_user  # Associate current user with the request object
+        user_model = get_user_model()
+        request.user = user_model.objects.filter(tg_id=init_data.user.id).first()
+        if request.user is None:
+            raise PermissionDenied("Telegram user is not registered")
 
-        response = self.get_response(request)
-
-        # Code to be executed for each request/response after
-        # the view is called.
-
-        return response
+        return self.get_response(request)
 ```
 
-To use `TWAAuthorizationMiddleware`, add it to your `MIDDLEWARE` setting in `settings.py`:
+## Enable Middleware
+
+Add the middleware to `settings.py`:
+
 ```python
 MIDDLEWARE = [
-    # other middleware classes
-    'path.to.TWAAuthorizationMiddleware',
+    # Other middleware classes...
+    "path.to.TelegramWebAppAuthMiddleware",
 ]
 ```
+
+!!! tip "Keep authentication boundaries explicit"
+
+    Middleware is convenient for apps where every route is Telegram-protected. If only a few views need Telegram auth, use a decorator or per-view helper instead.
