+++ docs/guide/examples/fastapi.md
@@ -1,10 +1,12 @@
-## Using with FastAPI
-Let's create some useful stuff according to [OAuth2 tutorial](https://fastapi.tiangolo.com/advanced/security/oauth2-scopes/?h=auth).
+# FastAPI Integration
 
-File `auth.py`:
+FastAPI dependencies are a good fit for Telegram Mini App authentication. Create one dependency that validates `initData` and returns the current Telegram user.
+
+## Authentication Dependency
 
 ```python
 import http
+from datetime import timedelta
 
 from fastapi import Depends
 from fastapi import HTTPException
@@ -12,11 +14,12 @@ from fastapi.security.http import HTTPAuthorizationCredentials
 from fastapi.security.http import HTTPBase
 
 from telegram_webapp_auth.auth import TelegramAuthenticator
-from telegram_webapp_auth.auth import WebAppUser
 from telegram_webapp_auth.auth import generate_secret_key
+from telegram_webapp_auth.data import WebAppUser
+from telegram_webapp_auth.errors import ExpiredInitDataError
 from telegram_webapp_auth.errors import InvalidInitDataError
 
-from .config import TelegramBotSettings  # Telegram Bot configuration
+from .config import TelegramBotSettings
 
 telegram_authentication_schema = HTTPBase()
 
@@ -32,37 +35,38 @@ def get_current_user(
     telegram_authenticator: TelegramAuthenticator = Depends(get_telegram_authenticator),
 ) -> WebAppUser:
     try:
-        init_data = telegram_authenticator.validate(auth_cred.credentials)
-    except InvalidInitDataError:
+        init_data = telegram_authenticator.validate(
+            init_data=auth_cred.credentials,
+            expr_in=timedelta(minutes=5),
+        )
+    except ExpiredInitDataError as exc:
+        raise HTTPException(
+            status_code=http.HTTPStatus.UNAUTHORIZED,
+            detail="Telegram init data has expired.",
+        ) from exc
+    except InvalidInitDataError as exc:
         raise HTTPException(
             status_code=http.HTTPStatus.FORBIDDEN,
-            detail="Forbidden access.",
-        )
-    except Exception:
-        raise HTTPException(
-            status_code=http.HTTPStatus.INTERNAL_SERVER_ERROR,
-            detail="Internal error.",
-        )
-    
+            detail="Telegram init data is invalid.",
+        ) from exc
+
     if init_data.user is None:
         raise HTTPException(
             status_code=http.HTTPStatus.FORBIDDEN,
-            detail="Forbidden access.",
+            detail="Telegram user data is required.",
         )
 
     return init_data.user
 ```
 
-Finally, we can use it as usual.
-
-File `app.py`:
+## Protected Route
 
 ```python
 from fastapi import Depends
 from fastapi import FastAPI
 from pydantic import BaseModel
 
-from telegram_webapp_auth.auth import WebAppUser
+from telegram_webapp_auth.data import WebAppUser
 
 from .auth import get_current_user
 
@@ -78,8 +82,12 @@ async def send_message(
     message: Message,
     user: WebAppUser = Depends(get_current_user),
 ):
-    """
-    Some logic...
-    """
-    ...
+    return {
+        "telegram_user_id": user.id,
+        "message": message.text,
+    }
 ```
+
+!!! note "Header format"
+
+    `HTTPBase()` expects an `Authorization` header with a scheme and credentials. A common frontend format is `Authorization: TWA <initData>`. Adapt the dependency if your client sends init data in another header or request body.
