+++ docs/license.md
@@ -1,3 +1,7 @@
-```
+# License
+
+`telegram-webapp-auth` is distributed under the MIT License.
+
+```text
 {% include 'LICENSE' %}
 ```
