+++ docs/references/data.md
@@ -1,1 +1,7 @@
+# Data Models
+
+Validated Telegram data is returned as dataclasses from `telegram_webapp_auth.data`.
+
+Use these models for typed access to users, chats, and launch metadata after validation succeeds.
+
 ::: telegram_webapp_auth.data
